requirements.txt
anyio==3.5.0
asgiref==3.5.0
autopep8==1.6.0
cffi==1.15.0
click==8.0.4
cryptography==36.0.2
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
jwt==1.3.1
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
PyMySQL==1.0.2
python-dotenv==0.19.2
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
toml==0.10.2
typing-extensions==4.1.1
uvicorn==0.17.6
